chore(tsutils): rename to `utils`
