+++ src/rules/no-mocks-import.ts
@@ -1,6 +1,6 @@
 import { TSESTree } from '@typescript-eslint/experimental-utils';
 import { posix } from 'path';
-import { createRule, getStringValue, isStringNode } from './tsUtils';
+import { createRule, getStringValue, isStringNode } from './utils';
 
 const mocksDirName = '__mocks__';
