+++ src/rules/prefer-inline-snapshots.ts
@@ -1,5 +1,5 @@
 import { AST_NODE_TYPES } from '@typescript-eslint/experimental-utils';
-import { createRule } from './tsUtils';
+import { createRule } from './utils';
 
 export default createRule({
   name: __filename,
