+++ src/rules/require-tothrow-message.ts
@@ -3,7 +3,7 @@ import {
   createRule,
   isExpectCall,
   parseExpectCall,
-} from './tsUtils';
+} from './utils';
 
 export default createRule({
   name: __filename,
