--- src/rules/util.js
@@ -1,2 +1,0 @@
-export const argument = node =>
-  node.parent.parent.arguments && node.parent.parent.arguments[0];
